% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{config}
\alias{config}
\title{Test config file}
\format{
A data frame with 2 columns and 7 rows
}
\source{
scROSHI: robust supervised hierarchical identification of single cells
\doi{10.1093/nargab/lqad058}
}
\usage{
config
}
\description{
Config file for the test_sce_data set
}
\details{
Config file to define major cell types and hierarchical subtypes.
It should be provided as a two-column data.frame where the first column are
the major cell types and the second column are the subtypes. If several
subtypes exists, they should be separated by comma.
}
\keyword{datasets}
