% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_my_correlation_test.R
\name{f_my_correlation_test}
\alias{f_my_correlation_test}
\title{Wilcox test}
\usage{
f_my_correlation_test(x, y, min_genes = 5)
}
\arguments{
\item{x}{Value 1}

\item{y}{Value 2}

\item{min_genes}{Minimum genes}
}
\description{
Calculate spearman correlation p-value or return preset values
}
\examples{
f_my_wilcox_test(rnorm(10,1,2),rnorm(10,5,2))
}
