% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_my_wilcox_test.R
\name{f_my_wilcox_test}
\alias{f_my_wilcox_test}
\title{Wilcox test}
\usage{
f_my_wilcox_test(x, y, min_genes = 5)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{y}{A numeric vector of values.}

\item{min_genes}{A numeric value defining the threshold for the minimum number of genes.}
}
\value{
A numeric value representing the p value of the Wilcox test.
}
\description{
Calculate Wilcox p-value or return preset values
}
\examples{
f_my_wilcox_test(rnorm(10, 1, 2), rnorm(10, 5, 2))
}
