% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_final_score.R
\name{f_plot_final_score}
\alias{f_plot_final_score}
\title{Score heatmap plot function}
\usage{
f_plot_final_score(final.score, major.type)
}
\arguments{
\item{final.score}{final score}

\item{major.type}{major type}
}
\value{

}
\description{
Score heatmap plot function
}
