% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_score_ctgenes_U.R
\name{f_score_ctgenes_U}
\alias{f_score_ctgenes_U}
\title{Calculate celltype score: U-test}
\usage{
f_score_ctgenes_U(sce, gset, min_genes = 5, verbose = 0)
}
\arguments{
\item{sce}{sce object}

\item{gset}{gene set}

\item{min_genes}{minimum number of genes}
}
\description{
Calculate celltype score: U-test
}
\examples{
#' \dontrun{
result <- f_score_ctgenes_U(sce, gset, min_genes = 5,verbose = 0)
}
}
