% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\docType{data}
\name{marker_list}
\alias{marker_list}
\title{Marker gene list for the test SCE data Set}
\format{
Marker gene list for all cell types as a list of genes with cell types as names
}
\source{
scROSHI - robust supervised hierarchical identification of single cells
\url{https://www.biorxiv.org/content/10.1101/2022.04.05.487176v1}
}
\usage{
marker_list
}
\description{
Marker gene list for the test_sce_data set
}
\keyword{datasets}
