% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_sce_data}
\alias{test_sce_data}
\title{Test SCE Data Set}
\format{
A SingleCellExperiment object with 3368 genes and 1316 cells
}
\source{
scROSHI: robust supervised hierarchical identification of single cells
\doi{10.1093/nargab/lqad058}
}
\usage{
test_sce_data
}
\description{
Data from a peripheral blood mononuclear cell experiments from an adult human.
}
\keyword{datasets}
