% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_annot_ctgenes.R
\name{f_annot_ctgenes}
\alias{f_annot_ctgenes}
\title{Choose best matching celltype}
\usage{
f_annot_ctgenes(m.cts, unknown, uncertain)
}
\arguments{
\item{m.cts}{celltype score}

\item{unknown}{threshold unknown}

\item{uncertain}{threshold uncertain}
}
\value{

}
\description{
Choose best matching celltype
}
