% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_annot_ctgenes.R
\name{f_annot_ctgenes}
\alias{f_annot_ctgenes}
\title{Choose best matching cell type}
\usage{
f_annot_ctgenes(m.cts, unknown, uncertain)
}
\arguments{
\item{m.cts}{Cell type score}

\item{unknown}{Threshold unknown}

\item{uncertain}{Threshold uncertain}
}
\description{
Choose best matching cell type
}
\examples{
\dontrun{
result <- f_annot_ctgenes(m.cts, unknown, uncertain)
}
}
