% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_my_wilcox_test.R
\name{f_my_wilcox_test}
\alias{f_my_wilcox_test}
\title{Wilcox test}
\usage{
f_my_wilcox_test(x, y, min_genes = 5)
}
\arguments{
\item{x}{value 1}

\item{y}{value 2}

\item{min_genes}{minimum genes}
}
\value{

}
\description{
Calculate wilcox p-value or return preset values
}
\examples{
f_my_wilcox_test(rnorm(10,1,2),rnorm(10,5,2))
}
