% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scROSHI-package.R
\docType{package}
\name{scROSHI}
\alias{scROSHI}
\title{Robust supervised hierarchical identification of single cells}
\description{
Identifying cell types based on expression profiles is a pillar of single cell analysis.
}
